requirements.txt
langchain
langchain-openai
langchain-community
yfinance
pandas
python-dotenv
streamlit